pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "floodrisk-hand"
version = "0.1.0"
description = "HAND-FIM flood inundation mapping with social vulnerability assessment"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Sebastian R.O. Marshall", email = "xec3cr@virginia.edu"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Hydrology",
]
dependencies = [
    "numpy>=1.24",
    "xarray>=2023.1",
    "rioxarray>=0.14",
    "geopandas>=0.13",
    "folium>=0.14",
    "requests>=2.31",
    "pyyaml>=6.0",
    "tqdm>=4.65",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov", "ruff"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py39"
